+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add filter icon to filters active info node
 - Fix missing context flag for show export button
 - Implement a sub tag regex (based on work by[calw20](https://github.com/calw20))
+- Support opening a URL by clicking on the sub tag in the tree
 
 ## v0.0.196 - 2021-01-21